pyproject.toml
[project]
name = "policyparser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.136.0",
    "jinja2>=3.1.6",
    "langchain-core>=1.3.0",
    "langchain-qwq>=0.3.4",
    "langgraph>=1.1.8",
    "loguru>=0.7.3",
    "markdownify>=1.2.2",
    "openai>=2.32.0",
    "pydantic>=2.13.2",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "starlette>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.11",
    "ty>=0.0.32",
]

[tool.ruff]
